Fix major bug with library inclusion in BuildQt56DevArchive.py
When searching for deployed DLLs, the DLLs are found and stored
in found_dlls. However, a mistake was made with post-processing said
DLLs - the Qt5 prefix was removed from elements that started with Qt5.
Later, the list of DLLs is simply appended to the list of libraries to
include, resulting in library name mismatches, e.g. "Svg" vs "Qt5Svg"
(latter is correct). Because of the mismatched names, the important
libraries found were not included within the archived binary
distribution.

This was fixed by removing the invalid prefix post-processing.